Added some caching of downloaded images.

+++ trunk/lib/PlokView.py
@@ -23,9 +23,10 @@ class PlokView(Base.View):
         self.ploklist = []
     
     def get_ploklist(self):
-        params = {'operation': 'get_ploks_newest'}
+        params = {'operation': 'get_ploks_newest', 'thumb_size' : '44'}
         if len(self.ploklist) > 0:
             params["lastid"] = self.ploklist[0]["id"]
+        # TODO: really use this lastid stuff
         params = urllib.urlencode({'operation': 'get_ploks_newest'})
         url = "http://www.plok.in/api/test.php"
         ip = appuifw.InfoPopup()
@@ -34,42 +35,47 @@ class PlokView(Base.View):
         data = simplejson.loads(f.read())
         f.close()
         self.ploklist = data["ploklist"]
-        ip.hide()
+        # ip.hide()
         ip.show(u"Dumping images", (50, 50), 60000, 100, appuifw.EHLeftVTop)
         for plok in self.ploklist:
-            imagefile = "D:\\%(id)s.jpg" % plok
-            # TODO: if not os.path.isfile(imagefile):
-            f = open(imagefile, "wb")
-            f.write(base64.decodestring(plok["image_base64"]))
-            f.close()
+            filename = "D:\\icon-%(id)s.jpg" % plok
+            if not os.path.isfile(filename):
+                f = open(filename, "wb")
+                f.write(base64.decodestring(plok["image_base64"]))
+                f.close()
         ip.hide()
 
     def get_plok(self, id):
-        params = urllib.urlencode({'operation': 'get_plok', 'id' : id})
+        params = urllib.urlencode({'operation': 'get_plok', 
+                                   'id' : id, 
+                                   'image_size' : '240'})
         url = "http://www.plok.in/api/test.php"
         ip = appuifw.InfoPopup()
         ip.show(u"Loading plok...", (50, 50), 60000, 100, appuifw.EHLeftVTop)
-        filename = u"D:\\plok.jpg"
-        urllib.urlretrieve("%s?%s" % (url, params), filename)
+        filename = u"D:\\plok-%s.jpg" % id
+        if not os.path.isfile(filename):
+            urllib.urlretrieve("%s?%s" % (url, params), filename)
         ip.hide()
         lock=e32.Ao_lock()
         content_handler = appuifw.Content_handler(lock.signal)
         content_handler.open(filename)
         lock.wait()
-        
 
+    def update_ploklist(self):
+        self.get_ploklist()
+        self.fill_items()
+        self.listbox.redraw_list()
+        
     def fill_items(self):
         # FIXME: check that these keys exist
         pattern = u"%(sender)s %(isotime)s\n%(title)s"
         self.items = [ pattern % (plok) for plok in self.ploklist ]
-        self.images = [ "D:\\%(id)s.jpg" % plok for plok in self.ploklist ]
+        self.images = [ "D:\\icon-%(id)s.jpg" % plok for plok in self.ploklist ]
 
     def item_selected(self):
         item = self.listbox.current()
         plok = self.ploklist[item]
         self.get_plok(plok["id"])
-        #print plok
-        #note(u"Selecting plok %s is not implemented yet, sorry." % plok["id"],"info")
 
     def handle_close(self):
         self.active = False
@@ -78,7 +84,7 @@ class PlokView(Base.View):
     def activate(self):
         self.active = True
         appuifw.app.screen = "full"
-        appuifw.app.menu = [(u"Update list", self.get_ploklist),
+        appuifw.app.menu = [(u"Update list", self.update_ploklist),
                             (u"Quit", self.close)]
         appuifw.app.exit_key_handler = self.handle_close
         if len(self.ploklist) == 0:
